flask/requirements.txt
backports.functools-lru-cache==1.6.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==2.2.4
mido==1.2.9
numpy==1.16.6
Pillow==6.2.2
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
subprocess32==3.5.4
urllib3==1.25.7
uWSGI==2.0.18
Werkzeug==0.16.0
